requirements.txt
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
cinemagoer==2022.12.27
greenlet==2.0.2
idna==3.4
IMDbPY==2022.7.9
lxml==4.9.2
python-dotenv==0.21.1
requests==2.28.2
soupsieve==2.3.2.post1
SQLAlchemy==2.0.2
typing_extensions==4.4.0
urllib3==1.26.14
wikipedia==1.4.0
